# lib/api-zod/src/generated/types/lineProbability.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface LineProbability {
  line: number;
  /**
     * @minimum 0
     * @maximum 1
     */
  over: number;
}