# lib/api-zod/src/generated/types/matchModel.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LineProbability } from './lineProbability';
import type { MatchModelDataQuality } from './matchModelDataQuality';
import type { MatchModelExpectedGoals } from './matchModelExpectedGoals';
import type { MatchModelResult } from './matchModelResult';
import type { MatchModelTeamStatsItem } from './matchModelTeamStatsItem';
import type { MatchModelTopScoresItem } from './matchModelTopScoresItem';

/**
 * Pure model reading of a scheduled match (not blended with the market).
 */
export interface MatchModel {
  modelVersion: string;
  expectedGoals: MatchModelExpectedGoals;
  result: MatchModelResult;
  /**
     * @minimum 0
     * @maximum 1
     */
  btts: number;
  topScores: MatchModelTopScoresItem[];
  goalLines: LineProbability[];
  teamStats: MatchModelTeamStatsItem[];
  dataQuality: MatchModelDataQuality;
}