# lib/api-zod/src/generated/types/valueBet.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LeagueCode } from './leagueCode';
import type { OddsQuote } from './oddsQuote';
import type { ValueBetConfidence } from './valueBetConfidence';

export type ValueBet = OddsQuote & ({
  fixtureId: string;
  league: LeagueCode;
  homeTeam: string;
  awayTeam: string;
  kickoff: Date;
  /**
     * @minimum 0
     * @maximum 1
     */
  modelProbability: number;
  /** @exclusiveMinimum 1 */
  fairOdds: number;
  expectedValuePct: number;
  modelVersion: string;
  /** @minimum 0 */
  sampleSize: number;
  confidence: ValueBetConfidence;
  /** @nullable */
  marketProbability?: number | null;
  /** @nullable */
  kellyFraction?: number | null;
  /** @nullable */
  bookmakersCount?: number | null;
});